pyproject.toml
[project]
name = "aris-voice-agent"
version = "1.0.0"
description = "Aris voice interface: browser-based voice agent with self-hosted TTS/STT on Modal GPU"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[whisper,openrouter,daily,silero,local-smart-turn]>=0.0.70",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0",
    "loguru>=0.7",
]
